Changed '__kernel_nlink_t' data type to match Linux/MIPS kernel type definition and to be consistent with the ABI. Done per conversation with Ralf (Linux/MIPS) maintainer.

+++ libc/sysdeps/linux/mips/bits/kernel_types.h
@@ -36,7 +36,7 @@ typedef long long      __kernel_loff_t;
 typedef unsigned int	__kernel_dev_t;
 typedef unsigned long	__kernel_ino_t;
 typedef unsigned int	__kernel_mode_t;
-typedef int		__kernel_nlink_t;
+typedef unsigned long	__kernel_nlink_t;
 typedef long		__kernel_off_t;
 typedef int		__kernel_pid_t;
 typedef int		__kernel_ipc_pid_t;
